Excluded frosted_api.h as dependance.
Applications can now build without kernel headers.

+++ newlib/libc/sys/frosted/ioctl.c
@@ -2,12 +2,11 @@
  * Frosted version of ioctl.
  */
 
-#include "frosted_api.h"
 #include "syscall_table.h"
 #include <errno.h>
-extern int sys_ioctl(int fd, const uint32_t cmd, void *arg);
+extern int sys_ioctl(int fd, const unsigned long cmd, void *arg);
 
-int ioctl (int fd, const uint32_t cmd, void *arg)
+int ioctl (int fd, const unsigned long cmd, void *arg)
 {
     int ret = sys_ioctl(fd, cmd, arg);
     if (ret < 0) {
