+++ newlib/libc/sys/frosted/umount.c
@@ -2,12 +2,11 @@
  * Frosted version of umount.
  */
 
-#include "frosted_api.h"
 #include "syscall_table.h"
 #include <errno.h>
-extern int sys_umount(char *target, uint32_t flags);
+extern int sys_umount(char *target, unsigned long flags);
 
-int umount(char *target, uint32_t flags)
+int umount(char *target, unsigned long flags)
 {
     int ret;
     (void)flags; /* flags unimplemented for now */
